requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.23
pymysql==1.1.1
python-dotenv==1.0.1
pydantic[email]==2.7.1
werkzeug==3.0.4
cryptography==43.0.1
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart
pytest==7.4.3
pytest-asyncio==0.21.1
mysql-connector-python==8.2.0
websockets==12.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
